layers: Move Query CoreChecks to own file
