+++ build-android/jni/Android.mk
@@ -76,6 +76,7 @@ LOCAL_SRC_FILES += $(SRC_DIR)/layers/convert_to_renderpass2.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/generated/layer_chassis_dispatch.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/generated/chassis.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/layer_options.cpp
+LOCAL_SRC_FILES += $(SRC_DIR)/layers/query_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/ray_tracing_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/wsi_validation.cpp
 LOCAL_SRC_FILES += $(SRC_DIR)/layers/xxhash.c
